Arrange rates menu to allow rate table selection
