+++ bin/i18n/json/app/en.json
@@ -3608,6 +3608,16 @@
       }
     },
     "rates_advanced": {
+      "advanced": {
+        "english": "Advanced",
+        "translation": "Advanced",
+        "needs_translation": false
+      },
+      "table": {
+        "english": "Rate Table",
+        "translation": "Rate Table",
+        "needs_translation": false
+      },
       "dyn_ceiling_gain": {
         "english": "Dynamic ceiling gain",
         "translation": "Dynamic ceiling gain",
