--- src/rfsuite/app/modules/rates_type/help.lua
@@ -1,15 +1,0 @@
---[[
-  Copyright (C) 2025 Rotorflight Project
-  GPLv3 — https://www.gnu.org/licenses/gpl-3.0.en.html
-]] --
-
-local rfsuite = require("rfsuite")
-local data = {}
-
-data['help'] = {}
-
-data['help']['default'] = {"@i18n(app.modules.rates_advanced.help_rate_table)@"}
-
-data['fields'] = {}
-
-return data
