--- src/rfsuite/app/modules/rates_type/init.lua
@@ -1,10 +1,0 @@
---[[
-  Copyright (C) 2025 Rotorflight Project
-  GPLv3 — https://www.gnu.org/licenses/gpl-3.0.en.html
-]] --
-
-local rfsuite = require("rfsuite")
-
-local init = {title = "@i18n(app.modules.rates_advanced.rates_type)@", section = "hidden", script = "rates_type.lua", image = "rates.png", order = 8, ethosversion = {1, 6, 2}}
-
-return init
