+++ src/rfsuite/i18n/en.json
@@ -5268,6 +5268,11 @@
           "needs_translation": false,
           "translation": "Accel"
         },
+        "advanced": {
+          "english": "Advanced",
+          "needs_translation": false,
+          "translation": "Advanced"
+        },
         "col": {
           "english": "Col",
           "needs_translation": false,
@@ -5392,6 +5397,11 @@
           "english": "Setpoint boost gain",
           "needs_translation": false,
           "translation": "Setpoint boost gain"
+        },
+        "table": {
+          "english": "Rate Table",
+          "needs_translation": false,
+          "translation": "Rate Table"
         },
         "yaw": {
           "english": "Yaw",
